# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Fiestamatic API - Dumaguete City fiesta navigator
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type CommunityPostPostType = typeof CommunityPostPostType[keyof typeof CommunityPostPostType];


export const CommunityPostPostType = {
  carpool: 'carpool',
  shared_table: 'shared_table',
  general: 'general',
} as const;

export interface CommunityPost {
  id: number;
  author_name: string;
  post_type: CommunityPostPostType;
  /** Target barangay for the fiesta visit */
  barangay: string;
  message: string;
  /** @nullable */
  contact_info?: string | null;
  /**
     * For carpool posts, number of available seats
     * @nullable
     */
  seats_available?: number | null;
  created_at: string;
}

export type CommunityPostInputPostType = typeof CommunityPostInputPostType[keyof typeof CommunityPostInputPostType];


export const CommunityPostInputPostType = {
  carpool: 'carpool',
  shared_table: 'shared_table',
  general: 'general',
} as const;

export interface CommunityPostInput {
  /**
     * @minLength 1
     * @maxLength 100
     */
  author_name: string;
  post_type: CommunityPostInputPostType;
  /** @minLength 1 */
  barangay: string;
  /**
     * @minLength 1
     * @maxLength 500
     */
  message: string;
  /** @maxLength 200 */
  contact_info?: string;
  /**
     * @minimum 1
     * @maximum 20
     */
  seats_available?: number;
}

export interface CommunitySummary {
  total_posts: number;
  carpool_posts: number;
  shared_table_posts: number;
  general_posts: number;
  /** Number of distinct barangays with posts */
  active_barangays: number;
}

export interface ErrorResponse {
  error: string;
}

export type GetCommunityPostsParams = {
/**
 * Filter by post type
 */
type?: GetCommunityPostsType;
};

export type GetCommunityPostsType = typeof GetCommunityPostsType[keyof typeof GetCommunityPostsType];


export const GetCommunityPostsType = {
  carpool: 'carpool',
  shared_table: 'shared_table',
  general: 'general',
} as const;