# lib/api-zod/src/generated/types/communityPost.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Fiestamatic API - Dumaguete City fiesta navigator
 * OpenAPI spec version: 0.1.0
 */
import type { CommunityPostPostType } from './communityPostPostType';

export interface CommunityPost {
  id: number;
  author_name: string;
  post_type: CommunityPostPostType;
  /** Target barangay for the fiesta visit */
  barangay: string;
  message: string;
  /** @nullable */
  contact_info?: string | null;
  /**
     * For carpool posts, number of available seats
     * @nullable
     */
  seats_available?: number | null;
  created_at: Date;
}