# lib/api-zod/src/generated/types/communityPostInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Fiestamatic API - Dumaguete City fiesta navigator
 * OpenAPI spec version: 0.1.0
 */
import type { CommunityPostInputPostType } from './communityPostInputPostType';

export interface CommunityPostInput {
  /**
     * @minLength 1
     * @maxLength 100
     */
  author_name: string;
  post_type: CommunityPostInputPostType;
  /** @minLength 1 */
  barangay: string;
  /**
     * @minLength 1
     * @maxLength 500
     */
  message: string;
  /** @maxLength 200 */
  contact_info?: string;
  /**
     * @minimum 1
     * @maximum 20
     */
  seats_available?: number;
}