# lib/api-zod/src/generated/types/communitySummary.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Fiestamatic API - Dumaguete City fiesta navigator
 * OpenAPI spec version: 0.1.0
 */

export interface CommunitySummary {
  total_posts: number;
  carpool_posts: number;
  shared_table_posts: number;
  general_posts: number;
  /** Number of distinct barangays with posts */
  active_barangays: number;
}